refactor: ensure correct imports
