+++ src/components/workout-crud/columns.tsx
@@ -12,7 +12,8 @@ import {
 	DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu"
 import { ColumnHeader } from "@/components/ui/data-table/column-header"
-import { InsertWorkoutSet } from "@/lib/db/schema/workouts/workout.schema"
+
+import { InsertWorkoutSet } from "@/lib/db/schema/workouts/schema"
 
 export const columns: ColumnDef<typeof InsertWorkoutSet._type>[] = [
 	{
