+++ src/lib/repositories/exercises/create.ts
@@ -1,7 +1,7 @@
 import { SqliteError } from "better-sqlite3"
 
 import { db } from "@/lib/db"
-import { exerciseMuscles } from "@/lib/db/schema/exercises/muscleGroups"
+import { exercisesOnMuscles } from "@/lib/db/schema/exercises/schema"
 import { exercises } from "@/lib/db/schema/exercises/schema"
 
 import { VCreateExercise } from "@/lib/repositories/validators"
@@ -17,7 +17,7 @@ export async function createExercises({
       .values(newExercise)
 
     const { changes } = await db
-      .insert(exerciseMuscles)
+      .insert(exercisesOnMuscles)
       .values(muscles.map(muscleID => ({
         exerciseId: exerciseID as number,
         muscleId: muscleID as number
