+++ src/lib/repositories/exercises/get.ts
@@ -3,7 +3,7 @@ import { db } from "@/lib/db"
 export async function getExercises() {
   return (await db.query.exercises.findMany({
     with: {
-      muscleGroups: {
+      exerciseMuscles: {
         columns: {
           exerciseId: false,
           muscleId: false
@@ -14,7 +14,7 @@ export async function getExercises() {
       }
     }
   })).map(v => ({
-    ...v, muscleGroups: v.muscleGroups.map(({ muscle: { id, name } }) => ({
+    ...v, muscleGroups: v.exerciseMuscles.map(({ muscle: { id, name } }) => ({
       id, name
     }))
   }))
