+++ src/lib/repositories/validators.ts
@@ -1,5 +1,5 @@
 import { z } from "zod"
-import { InsertWorkout, InsertWorkoutSet } from "@/lib/db/schema/workouts/workout.schema"
+import { InsertWorkout, InsertWorkoutSet } from "@/lib/db/schema/workouts/schema"
 import { InsertExercise } from "../db/schema/exercises/schema"
 
 export const VCreateWorkout = InsertWorkout.extend({
