+++ src/lib/repositories/workouts/controller.ts
@@ -1,0 +1,2 @@
+export * from "./create"
+export * from "./get"
